Worked flow for blog system with login register and moved to admin_app with changes in template for each header and sidebar

+++ admin_app/tasks.py
@@ -63,9 +63,13 @@ def Send_Bulk_Email(self,id):
 @shared_task
 def User_Registration_With_OTP(email):
     print("Inside celery with otp function",email)
-    user = User.objects.get(username=email)
-    user.delete()
-    user_obj=OTP_For_UserRegistration.objects.get(OTP_Email=email)
-    user_obj.delete()
+    try:
+        user = User.objects.get(username=email)
+        user_obj=OTP_For_UserRegistration.objects.get(OTP_Email=user)
+       
+        if user_obj ==False:
+            user.delete()
+    except:
+        return False
   
     return "DELETED"+str(user_obj)
